Matrix double to float

+++ src/main/utils/math/sigmoid.h
@@ -3,11 +3,11 @@
 /// @brief Computes the sigmoid activation of a given input.
 /// @param x The input value.
 /// @return The sigmoid of the input.
-double sigmoid(double x);
+float sigmoid(float x);
 
 /// @brief Computes the derivative of the sigmoid function at a given input.
 /// @param x The input value.
 /// @return The derivative of the sigmoid function at the input.
-double sigmoid_derivative(double x);
+float sigmoid_derivative(float x);
 
 #endif
